build: extract integration tests to separate app
