+++ examples/testapp1/gradle.properties
@@ -1,0 +1,5 @@
+greenmailVersion=2.1.2
+
+org.gradle.caching=true
+org.gradle.daemon=true
+org.gradle.parallel=true
