+++ examples/testapp1/grails-app/assets/javascripts/application.js
@@ -1,0 +1,20 @@
+// This is a manifest file that'll be compiled into application.js.
+//
+// Any JavaScript file within this directory can be referenced here using a relative path.
+//
+// You're free to add application-wide JavaScript to this file, but it's generally better
+// to create separate JavaScript files as needed.
+//
+//= require webjars/jquery/3.7.1/dist/jquery.js
+//= require webjars/bootstrap/5.3.3/dist/js/bootstrap.bundle
+//= require_self
+
+if (typeof jQuery !== 'undefined') {
+    (function($) {
+        $('#spinner').ajaxStart(function() {
+            $(this).fadeIn();
+        }).ajaxStop(function() {
+            $(this).fadeOut();
+        });
+    })(jQuery);
+}
