+++ examples/testapp1/grails-app/assets/stylesheets/application.css
@@ -1,0 +1,14 @@
+/*
+* This is a manifest file that'll be compiled into application.css, which will include all the files
+* listed below.
+*
+* Any CSS file within this directory can be referenced here using a relative path.
+*
+* You're free to add application-wide styles to this file and they'll appear at the top of the
+* compiled file, but it's generally better to create a new file per style scope.
+*
+*= require webjars/bootstrap/5.3.3/dist/css/bootstrap
+*= require webjars/bootstrap-icons/1.11.3/font/bootstrap-icons
+*= require grails
+*= require_self
+*/
