+++ examples/testapp1/grails-app/assets/stylesheets/errors.css
@@ -1,0 +1,90 @@
+.filename {
+    font-style: italic;
+}
+
+.exceptionMessage {
+    margin: 10px;
+    border: 1px solid #000;
+    padding: 5px;
+    background-color: #E9E9E9;
+}
+
+.stack,
+.snippet {
+    margin: 10px 0;
+}
+
+.stack,
+.snippet {
+    border: 1px solid #ccc;
+}
+
+/* error details */
+.error-details {
+    border: 1px solid #FFAAAA;
+    background-color:#FFF3F3;
+    line-height: 1.5;
+    overflow: hidden;
+    padding: 10px 0 5px 25px;
+}
+
+.error-details dt {
+    clear: left;
+    float: left;
+    font-weight: bold;
+    margin-right: 5px;
+}
+
+.error-details dt:after {
+    content: ":";
+}
+
+.error-details dd {
+    display: block;
+}
+
+/* stack trace */
+.stack {
+    padding: 5px;
+    overflow: auto;
+    height: 300px;
+}
+
+/* code snippet */
+.snippet {
+    background-color: #fff;
+    font-family: monospace;
+}
+
+.snippet .line {
+    display: block;
+}
+
+.snippet .lineNumber {
+    background-color: #ddd;
+    color: #999;
+    display: inline-block;
+    margin-right: 5px;
+    padding: 0 3px;
+    text-align: right;
+    width: 3em;
+}
+
+.snippet .error {
+    background-color: #fff3f3;
+    font-weight: bold;
+}
+
+.snippet .error .lineNumber {
+    background-color: #faa;
+    color: #333;
+    font-weight: bold;
+}
+
+.snippet .line:first-child .lineNumber {
+    padding-top: 5px;
+}
+
+.snippet .line:last-child .lineNumber {
+    padding-bottom: 5px;
+}
