+++ examples/testapp1/grails-app/conf/spring/resources.groovy
@@ -1,0 +1,3 @@
+// Place your Spring DSL code here
+beans = {
+}
