+++ examples/testapp1/grails-app/i18n/messages.properties
