+++ examples/testapp1/grails-app/i18n/messages_fr.properties
