+++ examples/testapp1/grails-app/init/testapp1/Application.groovy
@@ -1,14 +1,12 @@
-package grails.plugins.mail
+package testapp1
 
 import grails.boot.GrailsApp
 import grails.boot.config.GrailsAutoConfiguration
-import grails.plugins.metadata.PluginSource
 import groovy.transform.CompileStatic
 
-@PluginSource
 @CompileStatic
 class Application extends GrailsAutoConfiguration {
     static void main(String[] args) {
-        GrailsApp.run(Application)
+        GrailsApp.run(Application, args)
     }
 }
