+++ examples/testapp1/settings.gradle
@@ -1,0 +1,34 @@
+plugins {
+    id 'com.gradle.develocity' version '3.19.1'
+    id 'com.gradle.common-custom-user-data-gradle-plugin' version '2.0.2'
+}
+
+def isCI = System.getenv('CI') != null
+def isLocal = !isCI
+def isAuthenticated = System.getenv('DEVELOCITY_ACCESS_KEY') != null
+
+develocity {
+    server = 'https://ge.grails.org'
+    buildScan {
+        tag('gpc')
+        tag('gpc/grails-mail')
+        publishing.onlyIf { isAuthenticated }
+        uploadInBackground = isLocal
+    }
+}
+
+buildCache {
+    local { enabled = isLocal }
+    remote(develocity.buildCache) {
+        push = isCI && isAuthenticated
+        enabled = true
+    }
+}
+
+rootProject.name = 'testapp1'
+
+includeBuild('../../../grails-mail') {
+    dependencySubstitution {
+        substitute module('io.github.gpc:grails-mail') using project(':')
+    }
+}
