+++ examples/testapp1/src/integration-test/resources/assets/grailslogo.png
