+++ examples/testapp1/src/integration-test/resources/logback-test.xml
@@ -13,7 +13,7 @@
 
     <root level="ERROR"><appender-ref ref="STDOUT"/></root>
 
-    <logger name="com.icegreen.greenmail" level="DEBUG" additivity="false"><appender-ref ref="STDOUT"/></logger>
-    <logger name="grails.plugins.mail" level="DEBUG" additivity="false"><appender-ref ref="STDOUT"/></logger>
+    <logger name="com.icegreen.greenmail" level="INFO" additivity="false"><appender-ref ref="STDOUT"/></logger>
+    <logger name="io.github.gpc:grails-mail" level="DEBUG" additivity="false"><appender-ref ref="STDOUT"/></logger>
 
 </configuration>
